+++ src/ui/mod.rs
@@ -11,6 +11,7 @@ pub mod dummy_data;
 pub mod equalizer_panel;
 pub mod folder_browser;
 pub mod header_bar;
+pub mod lastfm_settings;
 mod library_commands;
 pub mod objects;
 pub mod open_files;